+++ src/main/java/Controllers/Users.java
@@ -12,7 +12,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.util.Scanner;
 
-@Path("Users/")
+@Path("users/")
 public class Users{
     @GET
     @Path("read/")
@@ -45,14 +45,14 @@ public class Users{
     @Path("create/")
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces(MediaType.APPLICATION_JSON)
-    public static void createUsers(@FormDataParam("UserEmailAdd") String UserEmailAdd, @FormDataParam("UserNameAdd") String UserNameAdd, @FormDataParam("UserPassAdd") String UserPassAdd){
+    public static String createUsers(@FormDataParam("UserIDAdd") Integer UserIDAdd, @FormDataParam("UserEmailAdd") String UserEmailAdd, @FormDataParam("UserNameAdd") String UserNameAdd, @FormDataParam("UserPassAdd") String UserPassAdd){
         System.out.println("users/create/");
         JSONArray read = new JSONArray();
         try{
             if(UserEmailAdd == null || UserNameAdd == null || UserPassAdd == null){
                 throw new Exception("One or more data parameters are missing values in the HTTP request.");
             }
-            System.out.println("users/create UserID=" + UserID);
+            System.out.println("users/create UserID=" + UserIDAdd);
 
             PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Users (UserID, UserEmail, UserName, UserPass, UserAdmin) VALUES (?, ?, ?, ?, ?)");
 
@@ -64,26 +64,44 @@ public class Users{
                 ps.setBoolean(5, false);
             }
 
+            ps.setInt(1, UserIDAdd);
             ps.setString(2, UserEmailAdd);
             ps.setString(3, UserNameAdd);
             ps.setString(4, UserPassAdd);
-
             ps.executeUpdate();
-            read.toString();
+
+            return "{\"status\": \"OK\"}";
+            //read.toString();
 
         } catch (Exception e){
             System.out.println("Database error: " + e.getMessage());
+            return "{\"error\": \"Unable to update item, please see server console for more info.\"}";
         }
     }
 
-    public static void deleteUsers(int UserID){
+    @POST
+    @Path("delete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public static String deleteUsers(@FormDataParam("UserID") Integer UserID){
         try{
+            if (UserID == null){
+                throw new Exception("The ID data parameter is missing from the HTTP request.");
+            }
+            System.out.println("user/delete UserID=" + UserID);
+
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM Users WHERE UserID = ?");
             ps.setInt(1, UserID);
             ps.execute();
+
+            return "{\"status\": \"OK\"}";
+
         } catch (Exception e) {
             e.printStackTrace();
             System.out.println("Database error: " + e.getMessage());
+            return "{\"error\": \"Unable to update item, please see server console for more info.\"}";
         }
     }
+
+
 }
